+++ src/utils/utils.py
@@ -192,7 +192,7 @@ def capture_subprocess_output(subprocess_args, new_env=None, profileMode=False):
     global rocprof_args
     # Format command for debug messages, formatting for rocprofv1 and rocprofv2
     command = " ".join(rocprof_args)
-    console_debug("subprocess", "Running: " + command)
+    console_debug("subprocess", "Running: " + command + " " + " ".join(subprocess_args))
     # Start subprocess
     # bufsize = 1 means output is line buffered
     # universal_newlines = True is required for line buffering
@@ -625,6 +625,18 @@ def run_prof(
     if rocprof_cmd.endswith("v2"):
         # rocprofv2 has separate csv files for each process
         results_files = glob.glob(workload_dir + "/out/pmc_1/results_*.csv")
+
+        # Combine results into single CSV file
+        combined_results = pd.concat(
+            [pd.read_csv(f) for f in results_files], ignore_index=True
+        )
+
+        # Overwrite column to ensure unique IDs.
+        combined_results["Dispatch_ID"] = range(0, len(combined_results))
+
+        combined_results.to_csv(
+            workload_dir + "/out/pmc_1/results_" + fbase + ".csv", index=False
+        )
     elif rocprof_cmd.endswith("v3"):
         # rocprofv3 requires additional processing for each process
         results_files = process_rocprofv3_output(
@@ -639,17 +651,23 @@ def run_prof(
             process_kokkos_trace_output(workload_dir, fbase)
         # TODO: add hip trace output processing
 
-    # Combine results into single CSV file
-    combined_results = pd.concat(
-        [pd.read_csv(f) for f in results_files], ignore_index=True
-    )
+        # Combine results into single CSV file
+        if results_files:
+            combined_results = pd.concat(
+                [pd.read_csv(f) for f in results_files], ignore_index=True
+            )
+        else:
+            console_warning(
+                f"Cannot write results for {fbase}.csv due to no counter csv files generated."
+            )
+            return
 
-    # Overwrite column to ensure unique IDs.
-    combined_results["Dispatch_ID"] = range(0, len(combined_results))
+        # Overwrite column to ensure unique IDs.
+        combined_results["Dispatch_ID"] = range(0, len(combined_results))
 
-    combined_results.to_csv(
-        workload_dir + "/out/pmc_1/results_" + fbase + ".csv", index=False
-    )
+        combined_results.to_csv(
+            workload_dir + "/out/pmc_1/results_" + fbase + ".csv", index=False
+        )
 
     if new_env:
         # flatten tcc for applicable mi300 input
@@ -750,7 +768,8 @@ def process_rocprofv3_output(rocprof_output, workload_dir, is_timestamps):
             )
         else:
             # when the input is not for timestamps, and counter csv file is not generated, we assume failed rocprof run and will completely bypass the file generation and merging for current pmc
-            console_error("No counter csv files generated, rocprofv3 run failed!!!")
+            results_files_csv = []
+            console_warning("No counter csv files generated, rocprofv3 run failed!!!")
 
     else:
         console_error("The output file of rocprofv3 can only support json or csv!!!")
@@ -810,7 +829,7 @@ def gen_sysinfo(
     df["workload_name"] = workload_name
 
     blocks = []
-    if ip_blocks == None:
+    if not ip_blocks:
         t = ["SQ", "LDS", "SQC", "TA", "TD", "TCP", "TCC", "SPI", "CPC", "CPF"]
         blocks += t
     else:
@@ -1120,10 +1139,10 @@ def print_status(msg):
 
 def set_locale_encoding():
     try:
-        locale.setlocale(locale.LC_ALL, "en_US.UTF-8")
+        locale.setlocale(locale.LC_ALL, "C.UTF-8")
     except locale.Error as error:
         console_error(
-            "Please ensure that the 'en_US.UTF-8' locale is available on your system.",
+            "Please ensure that the 'C.UTF-8' locale is available on your system.",
             exit=False,
         )
         console_error(error)
@@ -1249,3 +1268,16 @@ def merge_counters_spatial_multiplex(df_multi_index):
 
     final_df = pd.concat(result_dfs, keys=coll_levels, axis=1, copy=False)
     return final_df
+
+
+def convert_metric_id_to_panel_idx(metric_id):
+    # "4.02" -> 402
+    # "4.23" -> 423
+    # "4" -> 400
+    tokens = metric_id.split(".")
+    if len(tokens) == 1:
+        return int(tokens[0]) * 100
+    elif len(tokens) == 2:
+        return int(tokens[0]) * 100 + int(tokens[1])
+    else:
+        raise Exception(f"Invalid metric id: {metric_id}")